fix: harden search-after paging lifecycle
