+++ .agents/skills/foundatio-repositories/SKILL.md
@@ -137,9 +137,9 @@ IReadOnlyRepository<T>
 - **`FieldNot` is AND-NOT**: Multiple conditions inside `FieldNot` mean NOT A AND NOT B. For NOT (A AND B), nest `FieldAnd` inside `FieldNot`.
 - **Range operators + time-series indexes**: `FieldLessThanOrEqual(f => f.CreatedUtc, now)` does NOT narrow which daily/monthly indexes are queried. Always pair with `.Index(start, end)`.
 - **`FieldEquals` on analyzed text fields throws**: If the field has no `.keyword` sub-field, `FieldEquals` throws `QueryValidationException`. Use `FieldContains` for full-text search.
-- **v8+ Live `SearchAfterPaging()` requires a sort**: The repository adds the model id's sort-safe mapped field (such as `id.sort`) when possible. If the model has no `IIdentity` or the index sets `HasSortableIdField = false`, Live mode requires an explicit stable, unique sort and throws `QueryValidationException` when none exists. Point-in-time mode always makes Elasticsearch's implicit `_shard_doc` tiebreaker explicit after caller sorts, keeping the complete sort tuple reversible for backward cursors.
-- **Search-after direction is exclusive and disabling resets the session**: Setting an after cursor clears any before cursor, and vice versa. `SearchAfterPaging(false)` clears cursors, mode, and stored point-in-time state; close an abandoned point-in-time result before disabling or let its keep-alive expire.
-- **Externally-managed daily/monthly indexes are read-only adapters**: Use `ElasticReadOnlyRepositoryBase<T>`. Write methods can call `EnsureIndexAsync()` and create library-owned versioned partitions. Queries without `.Index(...)` and large-range fallbacks require the external system to maintain the `{Name}` umbrella alias. Override `GetIndexMappingFilter()` and `GetIndexDate()` together; the filter is authoritative and malformed wildcard matches returning `DateTime.MaxValue` are ignored.
+- **v8+ Live `SearchAfterPaging()` requires a sort**: The repository adds the model id's sort-safe mapped field (such as `id.sort`) when possible. If the model has no `IIdentity` or the index sets `HasSortableIdField = false`, Live mode requires an explicit stable, unique sort and throws `QueryValidationException` when none exists. Point-in-time mode makes Elasticsearch's implicit `_shard_doc` tiebreaker explicit after caller sorts, keeping the complete sort tuple reversible for backward cursors.
+- **Search-after sessions have strict invariants**: Direction is exclusive, cursor value count must match the final sort count, and snapshot paging cannot be mixed with PIT search-after. Raw null/empty/all-null arrays clear the cursor; decoded tokens preserve null sort values. Changing modes clears cursor/PIT/warning state, while reapplying the same mode preserves it. Scalar `FindOneAsync`, `CountAsync`, and query-based `ExistsAsync` reject PIT search-after options. Close caller-owned PITs when abandoning traversal; repository-owned PITs are closed on completion or failure.
+- **Externally-managed daily/monthly indexes are read-only adapters**: Use `ElasticReadOnlyRepositoryBase<T>`. Write methods can call `EnsureIndexAsync()` and create library-owned versioned partitions. Queries without `.Index(...)` and large-range fallbacks require the external system to maintain the `{Name}` umbrella alias. Override `MappingIndexPattern` and `GetIndexDate()` together; the pattern is authoritative and malformed wildcard matches returning `DateTime.MaxValue` are ignored.
 - **`PatchAsync(Ids, ...)` requires `Ids` type**: Use `new Ids(id1, id2)` -- `string[]` does not implicitly convert to `Ids`.
 - **`PatchAllAsync` with filter-only queries sends type-level notifications**: `EntityChanged` message has `Id = null`. Subscribers that need specific document IDs should re-query.
 - **Patch `DocumentsChanged` event has empty document list**: Only single-document `ActionPatch` populates `args.Documents`. All other patch types have an empty list.
