+++ .agents/skills/foundatio-repositories/references/patterns.md
@@ -146,9 +146,9 @@ foreach (var bucket in result.Aggregations.DateHistogram("date_created_utc").Buc
 
 Always use `SearchAfterPaging()` for deep pagination. Never use offset-based `.PageNumber()` for large result sets.
 
-The repository appends the id's sort-safe mapped field (for example `id.sort`) as a tiebreaker when the model implements `IIdentity` and the index guarantees that field is sortable. Otherwise, Live mode requires an explicit stable, unique sort and throws `QueryValidationException` when none exists. Point-in-time mode always makes Elasticsearch's implicit `_shard_doc` tiebreaker explicit after caller sorts, keeping the complete sort tuple reversible for backward cursors.
+The repository appends the id's sort-safe mapped field (for example `id.sort`) as a tiebreaker when the model implements `IIdentity` and the index guarantees that field is sortable. Otherwise, Live mode requires an explicit stable, unique sort and throws `QueryValidationException` when none exists. [Elasticsearch adds `_shard_doc` as an implicit PIT tiebreaker](https://www.elastic.co/docs/reference/elasticsearch/rest-apis/paginate-search-results#search-after); point-in-time mode makes it explicit after caller sorts so the complete tuple can be reversed for backward cursors.
 
-Forward and backward cursors are mutually exclusive: setting `SearchAfter(...)` or `SearchAfterToken(...)` clears `SearchBefore`, and vice versa. `SearchAfterPaging(false)` clears the complete local paging session, including cursors and stored point-in-time state. Close an abandoned point-in-time result with `ISupportPointInTime.ClosePointInTimeAsync(results)` before disabling; otherwise it remains open until its keep-alive expires.
+Forward and backward cursors are mutually exclusive and the cursor value count must equal the final sort count. Raw null/empty/all-null arrays clear the cursor; token-decoded arrays preserve null sort values. Changing Live/PIT mode clears the paging session; reapplying the same mode preserves it. Snapshot paging and PIT search-after are mutually exclusive, and scalar query operations reject PIT search-after. `SearchAfterPaging(false)` clears the complete local session. Close an abandoned caller-owned PIT with `ISupportPointInTime.ClosePointInTimeAsync(results)`; repository-owned PITs close on completion or failure.
 
 ```csharp
 var results = await repository.GetAllAsync(o => o.SearchAfterPaging().PageLimit(500));
