+++ src/Foundatio.Repositories.Elasticsearch/Repositories/ElasticReadOnlyRepositoryBase.cs
@@ -394,10 +394,15 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
 
     private static PagingStrategy GetPagingStrategy(ICommandOptions options)
     {
-        if (options.ShouldUseSearchAfterPagingPointInTime())
+        bool usePointInTime = options.ShouldUseSearchAfterPagingPointInTime();
+        bool useSnapshot = options.ShouldUseSnapshotPaging();
+        if (usePointInTime && useSnapshot)
+            throw new QueryValidationException("Snapshot paging and point-in-time search-after paging cannot be used together.");
+
+        if (usePointInTime)
             return PagingStrategy.SearchAfterPointInTime;
 
-        if (options.ShouldUseSnapshotPaging())
+        if (useSnapshot)
             return PagingStrategy.Snapshot;
 
         return PagingStrategy.Normal;
@@ -427,95 +432,110 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
             }
         }
 
-        if (options.HasAsyncQueryId())
+        try
         {
-            var queryId = options.GetAsyncQueryId();
-            ArgumentException.ThrowIfNullOrEmpty(queryId);
-
-            var response = await _client.AsyncSearch.GetAsync<TResult>(queryId, s =>
+            if (options.HasAsyncQueryId())
             {
-                if (options.HasAsyncQueryWaitTime())
-                    s.WaitForCompletionTimeout(options.GetAsyncQueryWaitTime());
-            }).AnyContext();
+                var queryId = options.GetAsyncQueryId();
+                ArgumentException.ThrowIfNullOrEmpty(queryId);
 
-            if (options.ShouldAutoDeleteAsyncQuery() && !response.IsRunning)
-                await RemoveQueryAsync(queryId).AnyContext();
+                var response = await _client.AsyncSearch.GetAsync<TResult>(queryId, s =>
+                {
+                    if (options.HasAsyncQueryWaitTime())
+                        s.WaitForCompletionTimeout(options.GetAsyncQueryWaitTime());
+                }).AnyContext();
 
-            _logger.LogRequest(response, options.GetQueryLogLevel());
-            if (!response.IsValidResponse && response.ApiCallDetails.HttpStatusCode.GetValueOrDefault() == 404)
-                throw new AsyncQueryNotFoundException(queryId);
+                if (options.ShouldAutoDeleteAsyncQuery() && !response.IsRunning)
+                    await RemoveQueryAsync(queryId).AnyContext();
 
-            result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
-        }
-        else if (options.HasSnapshotScrollId())
-        {
-            var scrollRequest = new ScrollRequest(options.GetSnapshotScrollId()!) { Scroll = options.GetSnapshotLifetime() };
-            var response = await _client.ScrollAsync<TResult>(scrollRequest).AnyContext();
-            _logger.LogRequest(response, options.GetQueryLogLevel());
-            result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
-        }
-        else
-        {
-            var searchDescriptor = await CreateSearchDescriptorAsync(query, options).AnyContext();
-            switch (pagingStrategy)
-            {
-                case PagingStrategy.Snapshot:
-                    searchDescriptor.Scroll(options.GetSnapshotLifetime());
-                    break;
-                case PagingStrategy.SearchAfterPointInTime:
-                    await ConfigurePointInTimeAsync(searchDescriptor, query, options).AnyContext();
-                    break;
+                _logger.LogRequest(response, options.GetQueryLogLevel());
+                if (!response.IsValidResponse && response.ApiCallDetails.HttpStatusCode.GetValueOrDefault() == 404)
+                    throw new AsyncQueryNotFoundException(queryId);
+
+                CapturePointInTimeId(options, response.Response.PitId);
+                result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
             }
-
-            if (query.ShouldOnlyHaveIds())
-                searchDescriptor.Source(false);
-
-            if (options.ShouldUseAsyncQuery())
+            else if (options.HasSnapshotScrollId())
             {
-                SearchRequest searchRequest = searchDescriptor;
-                var asyncSearchRequest = searchRequest.ToAsyncSearchSubmitRequest<TResult>();
-
-                if (options.HasAsyncQueryWaitTime())
-                    asyncSearchRequest.WaitForCompletionTimeout = options.GetAsyncQueryWaitTime();
-
-                var response = await _client.AsyncSearch.SubmitAsync<TResult>(asyncSearchRequest).AnyContext();
+                var scrollRequest = new ScrollRequest(options.GetSnapshotScrollId()!) { Scroll = options.GetSnapshotLifetime() };
+                var response = await _client.ScrollAsync<TResult>(scrollRequest).AnyContext();
                 _logger.LogRequest(response, options.GetQueryLogLevel());
                 result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
             }
             else
             {
-                var response = await _client.SearchAsync<TResult>(searchDescriptor).AnyContext();
-                _logger.LogRequest(response, options.GetQueryLogLevel());
-                result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
+                var searchDescriptor = await CreateSearchDescriptorAsync(query, options).AnyContext();
+                switch (pagingStrategy)
+                {
+                    case PagingStrategy.Snapshot:
+                        searchDescriptor.Scroll(options.GetSnapshotLifetime());
+                        break;
+                    case PagingStrategy.SearchAfterPointInTime:
+                        searchDescriptor.Indices(null);
+                        searchDescriptor.IgnoreUnavailable(null);
+                        await ConfigurePointInTimeAsync(searchDescriptor, query, options).AnyContext();
+                        break;
+                }
+
+                if (query.ShouldOnlyHaveIds())
+                    searchDescriptor.Source(false);
+
+                if (options.ShouldUseAsyncQuery())
+                {
+                    SearchRequest searchRequest = searchDescriptor;
+                    var asyncSearchRequest = searchRequest.ToAsyncSearchSubmitRequest<TResult>();
+
+                    if (options.HasAsyncQueryWaitTime())
+                        asyncSearchRequest.WaitForCompletionTimeout = options.GetAsyncQueryWaitTime();
+
+                    var response = await _client.AsyncSearch.SubmitAsync<TResult>(asyncSearchRequest).AnyContext();
+                    _logger.LogRequest(response, options.GetQueryLogLevel());
+                    CapturePointInTimeId(options, response.Response.PitId);
+                    result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
+                }
+                else
+                {
+                    var response = await _client.SearchAsync<TResult>(searchDescriptor).AnyContext();
+                    _logger.LogRequest(response, options.GetQueryLogLevel());
+                    CapturePointInTimeId(options, response.PitId);
+                    result = response.ToFindResults(options, ElasticIndex.Configuration.Serializer, _logger);
+                }
             }
-        }
 
-        await OnAfterQueryAsync(query, options, result).AnyContext();
+            await OnAfterQueryAsync(query, options, result).AnyContext();
 
-        if (pagingStrategy is PagingStrategy.Snapshot && !result.HasMore)
-        {
-            // clear the scroll
-            string? scrollId = result.GetScrollId();
-            if (!String.IsNullOrEmpty(scrollId))
+            if (pagingStrategy is PagingStrategy.Snapshot && !result.HasMore)
             {
-                var response = await _client.ClearScrollAsync(s => s.ScrollId(scrollId)).AnyContext();
-                _logger.LogRequest(response, options.GetQueryLogLevel());
+                // clear the scroll
+                string? scrollId = result.GetScrollId();
+                if (!String.IsNullOrEmpty(scrollId))
+                {
+                    var response = await _client.ClearScrollAsync(s => s.ScrollId(scrollId)).AnyContext();
+                    _logger.LogRequest(response, options.GetQueryLogLevel());
+                }
             }
-        }
 
-        if (pagingStrategy is PagingStrategy.SearchAfterPointInTime && !result.HasMore && options.IsRepoOwnedPointInTime())
+            if (pagingStrategy is PagingStrategy.SearchAfterPointInTime && !result.HasMore && options.IsRepoOwnedPointInTime())
+            {
+                bool closed = await ClosePointInTimeAsync(result).AnyContext();
+                if (!closed)
+                    _logger.LogWarning("Failed to close repository-owned point in time after paging completed; it will expire after its keep-alive window");
+                else
+                    options.DisableSearchAfterPaging();
+            }
+
+            if (allowCaching && !result.IsAsyncQueryRunning() && !result.IsAsyncQueryPartial())
+                await SetCachedQueryResultAsync(options, result, cacheSuffix: cacheSuffix).AnyContext();
+
+            ((IFindResults<TResult>)result).GetNextPageFunc = previousResults => GetNextPageFunc(previousResults, query, options);
+
+            return result;
+        }
+        catch
         {
-            bool closed = await ClosePointInTimeAsync(result).AnyContext();
-            if (!closed)
-                _logger.LogWarning("Failed to close repository-owned point in time after paging completed; it will expire after its keep-alive window");
+            await CloseRepositoryOwnedPointInTimeAfterFailureAsync(options).AnyContext();
+            throw;
         }
-
-        if (allowCaching && !result.IsAsyncQueryRunning() && !result.IsAsyncQueryPartial())
-            await SetCachedQueryResultAsync(options, result, cacheSuffix: cacheSuffix).AnyContext();
-
-        ((IFindResults<TResult>)result).GetNextPageFunc = previousResults => GetNextPageFunc(previousResults, query, options);
-
-        return result;
     }
 
     public async Task RemoveQueryAsync(string queryId)
@@ -566,6 +586,7 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
     public virtual async Task<FindHit<T>> FindOneAsync(IRepositoryQuery query, ICommandOptions? options = null)
     {
         options = ConfigureOptions(options?.As<T>());
+        ThrowIfPointInTimePagingIsUnsupported(options, nameof(FindOneAsync));
         if (IsCacheEnabled && (options.ShouldUseCache() || options.ShouldReadCache()) && !options.HasCacheKey())
             throw new ArgumentException("Cache key is required when enabling cache.", nameof(options));
 
@@ -606,6 +627,7 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
     public virtual async Task<CountResult> CountAsync(IRepositoryQuery query, ICommandOptions? options = null)
     {
         options = ConfigureOptions(options?.As<T>());
+        ThrowIfPointInTimePagingIsUnsupported(options, nameof(CountAsync));
 
         CountResult? result;
         if (IsCacheEnabled && options.ShouldReadCache())
@@ -678,6 +700,7 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
             throw new NotSupportedException("ExistsAsync requires the model type to implement IIdentity.");
 
         options = ConfigureOptions(options?.As<T>());
+        ThrowIfPointInTimePagingIsUnsupported(options, nameof(ExistsAsync));
         await OnBeforeQueryAsync(query, options, typeof(T)).AnyContext();
 
         await RefreshForConsistency(query, options).AnyContext();
@@ -836,6 +859,7 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
             // The repository opened this point-in-time, so it owns its lifecycle and must auto-close it once paging completes.
             // A caller-supplied point-in-time id is owned by the caller and is never auto-closed.
             pointInTimeId = await OpenPointInTimeAsync(query, options).AnyContext();
+            options.PointInTimeId(pointInTimeId);
             options.RepoOwnedPointInTime();
         }
 
@@ -881,10 +905,7 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
     {
         query = ConfigureQuery(query.As<T>()).Unwrap();
         string[] indices = ElasticIndex.GetIndexesByQuery(query);
-        // A point-in-time pins the target index set at the moment it was opened, so the request must not also set
-        // Indices(...) or IgnoreUnavailable() - doing so conflicts with the pinned context.
-        bool usePit = options.ShouldUseSearchAfterPagingPointInTime();
-        if (indices?.Length > 0 && !usePit)
+        if (indices?.Length > 0)
             search.Indices(String.Join(",", indices));
         if (HasVersion)
             search.SeqNoPrimaryTerm(HasVersion);
@@ -895,13 +916,43 @@ public abstract class ElasticReadOnlyRepositoryBase<T> : ISearchableReadOnlyRepo
             search.Timeout(timeout.ToElasticDuration());
         }
 
-        if (!usePit)
-            search.IgnoreUnavailable();
+        search.IgnoreUnavailable();
         search.TrackTotalHits(new TrackHits(options.ShouldTrackTotalHits()));
 
         await ElasticIndex.QueryBuilder.ConfigureSearchAsync(query, options, search).AnyContext();
 
         return search;
+    }
+
+    private static void CapturePointInTimeId(ICommandOptions options, string? pointInTimeId)
+    {
+        if (options.ShouldUseSearchAfterPagingPointInTime() && !String.IsNullOrEmpty(pointInTimeId))
+            options.PointInTimeId(pointInTimeId);
+    }
+
+    private async Task CloseRepositoryOwnedPointInTimeAfterFailureAsync(ICommandOptions options)
+    {
+        if (!options.IsRepoOwnedPointInTime() || !options.HasPointInTimeId())
+            return;
+
+        try
+        {
+            bool closed = await ClosePointInTimeAsync(options.GetPointInTimeId()).AnyContext();
+            if (closed)
+                options.DisableSearchAfterPaging();
+            else
+                _logger.LogWarning("Failed to close repository-owned point in time after a paging failure; it will expire after its keep-alive window");
+        }
+        catch (Exception ex)
+        {
+            _logger.LogWarning(ex, "Failed to close repository-owned point in time after a paging failure; it will expire after its keep-alive window");
+        }
+    }
+
+    private static void ThrowIfPointInTimePagingIsUnsupported(ICommandOptions options, string operation)
+    {
+        if (options.ShouldUseSearchAfterPagingPointInTime())
+            throw new QueryValidationException($"{operation} does not support point-in-time search-after paging. Use FindAsync or FindAsAsync so updated cursor and point-in-time state can be returned.");
     }
 
     protected virtual ICommandOptions<T> ConfigureOptions(ICommandOptions<T>? options)
